+++ abstutil/src/time.rs
@@ -170,7 +170,7 @@ impl<'a> Timer<'a> {
         }
         #[cfg(target_arch = "wasm32")]
         {
-            stdweb::console!(log, "%s", &line);
+            debug!("{}", line);
         }
         if let Some(ref mut sink) = maybe_sink {
             sink.println(line);
