+++ game/index.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <script type="module">
+        import { default as init } from './game.js';
+
+        async function run() {
+            await init('./game_bg.wasm');
+        }
+
+        run();
+    </script>
+</head>
+<body>
+        <div id="container">
+            <canvas id="canvas" width="1800" height="1000"></canvas>
+        </div>
+</body>
+<html>
