+++ widgetry_demo/index.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <script type="module">
+        import { default as init } from './widgetry_demo.js';
+
+        async function run() {
+            await init('./widgetry_demo_bg.wasm');
+        }
+
+        run();
+    </script>
+</head>
+<body>
+        <div id="container">
+            <canvas id="canvas" width="1800" height="1000"></canvas>
+        </div>
+</body>
+<html>
