+++ widgetry_demo/run_web.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+set -e
+wasm-pack build --dev --target web -- --no-default-features --features wasm
+cp index.html pkg
+cd pkg
+python3 -m http.server 8000
