+++ widgetry_demo/src/main.rs
@@ -1,0 +1,3 @@
+fn main() {
+    widgetry_demo::main();
+}
